Reorganising folder structure to make more sense
